refactor: 🔨 auth 관련 기능
