+++ .prettierrc.json
@@ -6,7 +6,7 @@
   "useTabs": false,
   "semi": true,
   "quoteProps": "as-needed",
-  "jsxSingleQuote": false,
+  "jsxSingleQuote": true,
   "trailingComma": "es5",
   "arrowParens": "always",
   "endOfLine": "lf",
