+++ src/auth/Login.tsx
@@ -1,8 +1,8 @@
 import { useForm, SubmitHandler } from 'react-hook-form';
 import axios from 'axios';
-import useLocalStorage from './localStorage';
+import useLocalStorage from './useLocalStorage';
 
-interface FormValue {
+interface IFormValue {
   email: string;
   password: string;
 }
@@ -15,20 +15,19 @@ const Login = () => {
     handleSubmit,
     setError,
     formState: { isSubmitting, errors },
-  } = useForm<FormValue>();
-  const [storedValue, setValue, removeValue] = useLocalStorage(TOKEN_KEY, '');
+  } = useForm<IFormValue>();
+  const [, setValue] = useLocalStorage(TOKEN_KEY, '');
 
-  const onSubmitHandler: SubmitHandler<FormValue> = async (data) => {
+  const onSubmitHandler: SubmitHandler<IFormValue> = async ({ email, password }) => {
     await axios
       .post('http://kdt.frontend.3rd.programmers.co.kr:5006/login', {
-        email: data.email,
-        password: data.password,
+        email,
+        password,
       })
       .then((res) => {
-        const token = res.data.token;
-        setValue(token);
+        setValue(res.data.token);
       })
-      .catch((err) => {
+      .catch(() => {
         setError(
           'password',
           { message: '아이디나 비밀번호 정보가 일치하지 않습니다. 다시 한번 확인해주세요.' },
@@ -38,41 +37,36 @@ const Login = () => {
   };
 
   return (
-    <>
-      <form onSubmit={handleSubmit(onSubmitHandler)} style={{ display: 'flex', flexDirection: 'column' }}>
-        <h1>로그인</h1>
-        <input
-          type="email"
-          placeholder="이메일을 입력해주세요"
-          {...register('email', {
-            required: '이메일 입력은 필수 입니다',
-            pattern: {
-              value: /\S+@\S+\.\S+/,
-              message: '이메일 형식에 맞지 않습니다.',
-            },
-          })}
-        />
-        <span>{errors?.email?.message}</span>
-        <input
-          type="password"
-          placeholder="비밀번호를 입력해주세요"
-          {...register('password', {
-            required: '비밀번호 입력은 필수 입니다',
-            minLength: {
-              value: 7,
-              message: '7자리 이상의 비밀번호를 입력해주세요',
-            },
-          })}
-        />
-        <span>{errors?.password?.message}</span>
-        <button type="submit" disabled={isSubmitting}>
-          로그인
-        </button>
-      </form>
-
-      <h1>로그아웃</h1>
-      <button onClick={removeValue}>로그아웃</button>
-    </>
+    <form onSubmit={handleSubmit(onSubmitHandler)} style={{ display: 'flex', flexDirection: 'column' }}>
+      <h1>로그인</h1>
+      <input
+        type='email'
+        placeholder='이메일을 입력해주세요'
+        {...register('email', {
+          required: '이메일 입력은 필수 입니다',
+          pattern: {
+            value: /\S+@\S+\.\S+/,
+            message: '이메일 형식에 맞지 않습니다.',
+          },
+        })}
+      />
+      <span>{errors?.email?.message}</span>
+      <input
+        type='password'
+        placeholder='비밀번호를 입력해주세요'
+        {...register('password', {
+          required: '비밀번호 입력은 필수 입니다',
+          minLength: {
+            value: 7,
+            message: '7자리 이상의 비밀번호를 입력해주세요',
+          },
+        })}
+      />
+      <span>{errors?.password?.message}</span>
+      <button type='submit' disabled={isSubmitting}>
+        로그인
+      </button>
+    </form>
   );
 };
