+++ src/auth/SignUp.tsx
@@ -2,7 +2,7 @@ import { useForm, SubmitHandler } from 'react-hook-form';
 import axios from 'axios';
 import { useRef } from 'react';
 
-interface FormValue {
+interface IFormValue {
   email: string;
   fullName: string;
   password: string;
@@ -15,16 +15,16 @@ const SignUp = () => {
     handleSubmit,
     watch,
     formState: { isSubmitting, errors },
-  } = useForm<FormValue>();
+  } = useForm<IFormValue>();
 
   const passwordRef = useRef<string | null>(null);
   passwordRef.current = watch('password');
 
-  const onSubmitHandler: SubmitHandler<FormValue> = async (data) => {
+  const onSubmitHandler: SubmitHandler<IFormValue> = async ({ email, fullName, password }) => {
     await axios.post('http://kdt.frontend.3rd.programmers.co.kr:5006/signup', {
-      email: data.email,
-      fullName: data.fullName,
-      password: data.password,
+      email,
+      fullName,
+      password,
     });
   };
 
@@ -32,8 +32,8 @@ const SignUp = () => {
     <form onSubmit={handleSubmit(onSubmitHandler)} style={{ display: 'flex', flexDirection: 'column' }}>
       <h1>회원가입</h1>
       <input
-        type="email"
-        placeholder="이메일을 입력해주세요"
+        type='email'
+        placeholder='이메일을 입력해주세요'
         {...register('email', {
           required: '이메일 입력은 필수 입니다',
           pattern: {
@@ -44,8 +44,8 @@ const SignUp = () => {
       />
       <span>{errors?.email?.message}</span>
       <input
-        type="text"
-        placeholder="이름을 입력해주세요!"
+        type='text'
+        placeholder='이름을 입력해주세요!'
         {...register('fullName', {
           required: '이름 입력은 필수 입니다',
           minLength: {
@@ -56,8 +56,8 @@ const SignUp = () => {
       />
       <span>{errors?.fullName?.message}</span>
       <input
-        type="password"
-        placeholder="비밀번호를 입력해주세요"
+        type='password'
+        placeholder='비밀번호를 입력해주세요'
         {...register('password', {
           required: '비밀번호 입력은 필수 입니다',
           minLength: {
@@ -68,15 +68,15 @@ const SignUp = () => {
       />
       <span>{errors?.password?.message}</span>
       <input
-        type="password"
-        placeholder="비밀번호를 한번 더 입력해주세요"
+        type='password'
+        placeholder='비밀번호를 한번 더 입력해주세요'
         {...register('passwordConfrim', {
           required: '비밀번호 확인 입력은 필수 입니다',
           validate: (value) => (value !== passwordRef.current ? '비밀번호가 일치하지 않습니다' : true),
         })}
       />
       <span>{errors?.passwordConfrim?.message}</span>
-      <button type="submit" disabled={isSubmitting}>
+      <button type='submit' disabled={isSubmitting}>
         회원가입
       </button>
     </form>
